fix(ticket): Fix load tickets

+++ app/Http/Controllers/Ticket/TicketIndexController.php
@@ -19,7 +19,7 @@ class TicketIndexController extends MainController
 
         $data['tickets'] = $ticket->getAllByCompanyId((int) Auth::user()->company_id, $search);
         $data['search'] = $search;
-        $data['customers'] = Customer::whereCompany((int) Auth::user()->company);
+        $data['customers'] = Customer::whereCompany(Auth::user()->company);
 
         return view('ticket.index', $data);
     }
